==> code/01_preprocessing.ipynb <==
In this section, we will investigate each dataset in order to better predict. 

### Import Libraries

import numpy as np
import pandas as pd
import re
import os
import pickle
from scipy.stats import shapiro, ttest_ind, chi2, mannwhitneyu

### Import data

holiday = pd.read_csv('../data/holidays_events.csv')
oil = pd.read_csv('../data/oil.csv')
store = pd.read_csv('../data/stores.csv')
transaction = pd.read_csv('../data/transactions.csv')
train = pd.read_csv('../data/train.csv')
test = pd.read_csv('../data/test.csv')

### Helper

def print_summary(data):
    """This function will print out the data summary within our data frame.
    It will print out the results including data type, mean, median, standard deviation, missing values and so on.
    
    Input:
    data -> dataframe 
    
    Output:
    summary -> statistical summary of the data frames"""
    
    for col, ctype in data.dtypes.items():
        missing = data[col].isnull().sum()
        summary = data[col].describe()
        total = len(data[col])
        print(f"Total number of rows: {total}")
        print(f"Missing value: {missing}")
        print(f"{col} summary: ")
        print(summary)
       
        print('-' * 50)
    

def create_new_path(path):
    
    """This function will create a new folder.
    
    Input: 
    path -> str (specify the path you want to create)
    """
    
    # try to create a new path in our data
    # if folders are already existed, then will get an error message
    try: 
        os.mkdir(path) 
    except OSError as e: 
        print(e)  

### Create new folders

output_path = '../output'
asset_path = '../asset'

for path in [output_path, asset_path]:
    create_new_path(path)

### Investigate

Since we have few different data frames, we need to investigate each data further. 

- holiday

holiday.head()

holiday.tail()

holiday.type.value_counts()

It looks like data description is in Spanish. Can we use description column later or?

holiday.transferred.value_counts()

There are some trasnferred holidays

holiday.info()

Relatively smaller. 350 rows total and looks like we do not have missing values. One thing to note is that the column date is in character values. It might be beneficial to change the column to datetime object. 

holiday.loc[:, 'date'] = pd.DatetimeIndex(holiday.date)

print_summary(holiday)

A quick data summary for the dataset. We may be able to join the date column. Might be some sort of seasonality. 

- oil

oil.head()

oil.tail()

oil.info()

oil.head()

oil.loc[:, 'date'] = pd.to_datetime(oil.date)

Still relatively small. The price of gas on a given date. Notice that the date column is also in character, so we need to convert it. 

oil.loc[:, 'date'] = pd.DatetimeIndex(oil.date)

print_summary(oil)

some missing value in this dataframe.

oil.fillna(method = 'ffill', inplace = True)

For the oil price, there are some missing valeus. Therefore, we can impute these missing values using the historical price. This might not be necessarily true, but it should not fluctuate that much, so for now we can use previous values. 

oil.fillna(method = 'bfill', inplace = True)

The first row has a missing value (1/1/13). Assuming this happened because of the new year, but let's use the price of 1/2/13 for this specific row. 

oil.to_csv('../output/complete_oil.csv', index = False)

- store

store.head()

store.tail()

store.type.value_counts()

store.cluster.value_counts().sort_index()

store.info()

store and cluster seem to be better with categorical values. need to change them into character

store.loc[:, 'store_nbr'] = store.store_nbr.apply(lambda x: str(x))
store.loc[:, 'cluster'] = store.cluster.apply(lambda x: str(x))

print_summary(store)

- transaction

transaction.head()

transaction.tail()

transaction.info()

non missing and not many columns. change date and store_sbr data type

transaction.head()

transaction.tail()

transaction.loc[:, 'date'] = pd.DatetimeIndex(transaction.date)
transaction.loc[:, 'store_nbr'] = transaction.store_nbr.apply(lambda x: str(x))

print_summary(transaction)

- train

train.head()

train.tail()

train.info()

for train, we have a good amount of data, but the number of columns is not that high. change their data type values accordingly.

check missing dates

train_check = train.set_index('date')

train_check.index = pd.to_datetime(train_check.index)

print(pd.date_range(start= min(train_check.index),
              end = max(train_check.index)).difference(train_check.index))

For Christmas, looks like all the stores are closed.

christmas_df = pd.DataFrame()
copy = train[train.date == '2013-01-01']

for christmas in ['2013-12-25', '2014-12-25', '2015-12-25', '2016-12-25']:
    copy.loc[:,'date'] = christmas
    christmas_df = christmas_df.append(copy, ignore_index= True)

train = pd.concat([train,  christmas_df]).drop(columns = 'id')

print_summary(train)

train.loc[:, 'date'] = pd.DatetimeIndex(train.date)
train.loc[:, 'store_nbr'] = train.store_nbr.apply(lambda x: str(x))

train.sort_values(by = ['date', 'store_nbr'], inplace = True)

- test

test.head()

test.tail()

test.info()

change the data types accordingly.

test.loc[:, 'date'] = pd.DatetimeIndex(test.date)
test.loc[:, 'store_nbr'] = test.store_nbr.apply(lambda x: str(x))

check missing dates

test_check = test.set_index('date')

test_check.index = pd.to_datetime(test_check.index)

print(pd.date_range(start= min(test_check.index),
              end = max(test_check.index)).difference(test_check.index))

print_summary(test)

### Combine data

After looking into data, we can combine the results

counts = pd.DataFrame(holiday.date.value_counts()).rename(columns = {'date':'holiday_counts'})

counts.loc[:, 'is_multiple'] = counts.holiday_counts.apply(lambda x: 1 if x > 1 else 0)

counts.reset_index(inplace = True)

counts = counts.rename(columns= {'index':'date'})

Let's use holiday counts rather the actual holiday dataframe since it causes duplicate issues.

holiday[holiday.transferred == True].to_pickle('../asset/transfer_holidays.pkl')

pickle transfer holidays

train_df = train.merge(oil, left_on= 'date', right_on = 'date', how = 'left').merge(counts, left_on= 'date', right_on= 'date', how = 'left').merge(store, left_on= 'store_nbr', right_on = 'store_nbr', how = 'left')

test_df =test.merge(oil, left_on= 'date', right_on = 'date', how = 'left').merge(counts, left_on= 'date', right_on= 'date', how = 'left').merge(store, left_on= 'store_nbr', right_on = 'store_nbr', how = 'left')

train_df.rename(columns= {'type_x':'holiday_type',
                          'type_y':'store_type'}, inplace = True)

test_df.rename(columns= {'type_x':'holiday_type',
                          'type_y':'store_type'}, inplace = True)

### Export data

train_df.to_pickle('../output/train_df.pkl')
test_df.to_pickle('../output/test_df.pkl')
