==> code/04_evaluation.ipynb <==
### Import libraries

import numpy as np
import pandas as pd
import pickle
from sklearn.metrics import mean_absolute_error
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### Helper

def calculate_total_errors(pred_dict, actual, yhat):
    
    """This function will calculate the erorrs using mean absolute erorrs.
    
    Input:
    pred_dict -> dict (key: unique_key, values: y and yhat)
    actual -> (values: the actual y value)
    yhat -> (values: the predicdted value from the model)
    
    Output:
    error -> int (mean absolute erorr)"""
    

    
    error_total = 0
    # looping through the dictionary and summing the error.
    for key in pred_dict:
        error = mean_absolute_error(pred_dict[key][actual], pred_dict[key][yhat])
        error_total += error
        
    return error_total

### Import data

- linear regression

with open('../asset/lr_model/lr_train_dict.pkl', 'rb') as f:
    lr_train_dict = pickle.load(f)

with open('../asset/lr_model/lr_test_dict.pkl', 'rb') as f:
    lr_test_dict = pickle.load(f)

- lasso

with open('../asset/lasso_model/lasso_train_dict.pkl', 'rb') as f:
    lasso_train_dict = pickle.load(f)

with open('../asset/lasso_model/lasso_test_dict.pkl', 'rb') as f:
    lasso_test_dict = pickle.load(f)

- xgb

with open('../asset/xgb_model/xgb_train_dict.pkl', 'rb') as f:
    xgb_train_dict = pickle.load(f)

with open('../asset/xgb_model/xgb_test_dict.pkl', 'rb') as f:
    xgb_test_dict = pickle.load(f)

- combined

with open('../asset/combined_model/total_train_dict.pkl', 'rb') as f:
    total_train_dict = pickle.load(f)

with open('../asset/combined_model/total_test_dict.pkl', 'rb') as f:
    total_test_dict = pickle.load(f)

- prophet

with open('../asset/prophet_model/combined_dict.pkl', 'rb') as f:
    prophet_combined_dict = pickle.load(f)

### Error Check

- lr

lr_train_error = calculate_total_errors(lr_train_dict, 'train_y', 'train_yhat')
lr_test_error = calculate_total_errors(lr_test_dict,'test_y', 'test_yhat' )

- lasso

lasso_train_error = calculate_total_errors(lasso_train_dict, 'train_y', 'train_yhat')
lasso_test_error = calculate_total_errors(lasso_test_dict, 'test_y', 'test_yhat')

- xgb

xgb_train_error = calculate_total_errors(xgb_train_dict, 'train_y', 'train_yhat')
xgb_test_error = calculate_total_errors(xgb_test_dict, 'test_y', 'test_yhat')

- combined

total_train_error = calculate_total_errors(total_train_dict, 'actual', 'yhat')
total_test_error = calculate_total_errors(total_test_dict, 'actual', 'yhat')

- prophet

train_error = 0
test_error = 0

for key in prophet_combined_dict:
    data = prophet_combined_dict[key]
    train = data[data.indicator == 'train']
    test = data[data.indicator == 'test']
    
    train_e = mean_absolute_error(train['y'], train['yhat'])
    test_e = mean_absolute_error(test['y'], test['yhat'])
    
    train_error += train_e
    test_error += test_e

print(f"Using Linear Regression, the mean absolute error in the training dataset is {round(lr_train_error, 4)}")
print(f"Using Linear Regression, the mean absolute error in the validation dataset is {round(lr_test_error, 4)}")
print('-' * 90)
print(f"Using Lasso, the mean absolute error in the training dataset is {round(lasso_train_error, 4)}")
print(f"Using Lasso, the mean absolute error in the validation dataset is {round(lasso_test_error, 4)}")
print('-' * 90)
print(f"Using XGBoost, the mean absolute error in the training dataset is {round(xgb_train_error, 4)}")
print(f"Using XGBoost, the mean absolute error in the validation dataset is {round(xgb_test_error, 4)}")
print('-' * 90)
print(f"Using Lasso & XGBoost, the mean absolute error in the training dataset is {round(total_train_error, 4)}")
print(f"Using Lasso & XGBoost, the mean absolute error in the validation dataset is {round(total_test_error, 4)}")
print('-' * 90)
print(f"Using Prophet, the mean absolute error in the training dataset is {round(train_error, 4)}")
print(f"Using Prophet, the mean absolute error in the validation dataset is {round(test_error, 4)}")

Looks like XGBoost has the best performance, but overfits. Let's hyper parameter tune the model.
